Refactor code for simplicity and improve IA of HEM drawer.
