+++ src/views/search/SummaryUx.tsx
@@ -1,28 +1,15 @@
-import {
-  VuiSpacer,
-  VuiTitle,
-  VuiHorizontalRule,
-  VuiSummary,
-  VuiFlexContainer,
-  VuiFlexItem,
-  VuiSpinner,
-  VuiText,
-  VuiButtonSecondary,
-} from "../../ui";
+import { VuiSpacer, VuiTitle, VuiHorizontalRule, VuiSummary } from "../../ui";
 import {
   sanitizeCitations,
   reorderCitations,
   applyCitationOrder,
 } from "../../ui/utils/citations";
-import { useHemScore } from "../../utils/useHemScore";
 import { useSearchContext } from "../../contexts/SearchContext";
 import { SearchResultList } from "./results/SearchResultList";
 import { ProgressReport } from "./progressReport/ProgressReport";
 import { SummaryCitation } from "./summary/SummaryCitation";
 import { DeserializedSearchResult } from "./types";
-import { ConfidenceBadge } from "../../utils/ConfidenceBadge";
-import { useState } from "react";
-import { HemDrawer } from "./controls/HemDrawer";
+import { ConfidenceScore } from "./summary/ConfidenceScore";
 
 export const SummaryUx = () => {
   const {
@@ -33,10 +20,7 @@ export const SummaryUx = () => {
     searchResultsRef,
     selectedSearchResultPosition,
     summaryEnableHem,
-    hfToken,
   } = useSearchContext();
-
-  const [isHemDrawerOpen, setIsHemDrawerOpen] = useState(false);
 
   const rawSummary = summarizationResponse?.summary[0]?.text;
   const unorderedSummary = sanitizeCitations(rawSummary);
@@ -53,24 +37,6 @@ export const SummaryUx = () => {
       );
     }
   }
-
-  const { isFetchingHemScore, confidenceLevel } = useHemScore(
-    summaryEnableHem,
-    hfToken,
-    rawSummary,
-    summarySearchResults
-  );
-  //    : { isFetchingHemScore: false, confidenceLevel: getConfidenceLevel(0) };
-
-  //  let isFetchingHemScore = false;
-  //  let confidenceLevel = getConfidenceLevel(0);
-  //  if (summaryEnableHem) {
-  //    ({ isFetchingHemScore, confidenceLevel } = useHemScore(
-  //      hfToken,
-  //      rawSummary,
-  //      summarySearchResults
-  //    ));
-  //  }
 
   return (
     <>
@@ -93,36 +59,10 @@ export const SummaryUx = () => {
           <VuiSpacer size="s" />
 
           {summaryEnableHem && (
-            <>
-              <VuiFlexContainer alignItems="center">
-                <VuiFlexItem grow={false} shrink={false}>
-                  {isFetchingHemScore ? (
-                    <VuiFlexContainer alignItems="center" spacing="xs">
-                      <VuiFlexItem grow={false}>
-                        <VuiSpinner size="xs" />
-                      </VuiFlexItem>
-
-                      <VuiFlexItem grow={false}>
-                        <VuiText size="s">
-                          <p>Evaluating confidence</p>
-                        </VuiText>
-                      </VuiFlexItem>
-                    </VuiFlexContainer>
-                  ) : (
-                    <ConfidenceBadge confidenceLevel={confidenceLevel} />
-                  )}
-                </VuiFlexItem>
-                <VuiFlexItem grow={false} shrink={false}>
-                  <VuiButtonSecondary
-                    color="subdued"
-                    size="s"
-                    onClick={() => setIsHemDrawerOpen(true)}
-                  >
-                    What's this?
-                  </VuiButtonSecondary>
-                </VuiFlexItem>
-              </VuiFlexContainer>
-            </>
+            <ConfidenceScore
+              rawSummary={rawSummary}
+              summarySearchResults={summarySearchResults}
+            />
           )}
 
           <VuiSpacer size="l" />
@@ -143,11 +83,6 @@ export const SummaryUx = () => {
             setSearchResultRef={(el: HTMLDivElement | null, index: number) =>
               (searchResultsRef.current[index] = el)
             }
-          />
-
-          <HemDrawer
-            isOpen={isHemDrawerOpen}
-            onClose={() => setIsHemDrawerOpen(false)}
           />
         </>
       )}
